requirements.txt
tornado>5.1,<5.2
py-healthcheck>1.7,<1.8
dependency_injector>3.14,<3.15
pyyaml>=3.13,<3.14
pymysql>0.9,<0.10